move request handler to its own class
